moved footer to its own directory
